Move twilio.api functions to their own file
